Move file from helpers to helper folder
